Proposal for addressing strings docs issue
See proposed change for details.

If we value `puppet strings` output for generating documentation, we
need to address a shortcoming it has with regards to data-in-modules.
This commit proposes a best practice that assumes the tooling as it
exists today.

Note that as an alternative, the tension/shortcoming could be resolved
by an improvement to puppet-strings. Discuss.